+++ src/main/scala/infcalcs/tables/ConstructedTable.scala
@@ -1,7 +1,7 @@
 package infcalcs.tables
 
 /** Class for building a contingency table from scratch. */
-class ConstructedTable(v: Vector[Vector[Int]]) extends ContTable {
+class ConstructedTable(v: Vector[Vector[Double]]) extends ContTable {
   lazy val rows = table.length
   lazy val cols = if (table.isEmpty) 0 else table(0).length
   lazy val table = v
